Add support for simple keyword replacement.
There have been some requests for the ability to do a keyword
replacement within a string.  This restricts that ability to do full
JSON replacements, so change the keyword mapping to support literal
replacements with ## instead of @@ for separation of simple and
json mappings.  This means there is one mappings object, but
each mapping can be used in either way.

Also added some better error reporting for bad JSON files.  The error
message was previously getting squashed and it wasn't printing the
post-keyword replacement JSON.  Now it prints the error message right,
and adds the post-replacement JSON.  So you can compare pre and post.